implementation of awesome api to obtain currency value in real time

+++ cambio-service/cambio-service/src/main/java/com/igorbarreto/cambioservice/model/Cambio.java
@@ -32,15 +32,15 @@ public class Cambio implements Serializable {
     private String to;
 
     @Column(nullable = false)
-    private BigDecimal conversionFactor;
+    private Double conversionFactor;
 
     @Transient
-    private BigDecimal convertedValue;
+    private Double convertedValue;
 
     @Transient
     private String environment;
 
-    public Cambio(Long id, String from, String to, BigDecimal conversionFactor, BigDecimal convertedValue, String environment) {
+    public Cambio(Long id, String from, String to, Double conversionFactor, Double convertedValue, String environment) {
         this.id = id;
         this.from = from;
         this.to = to;
@@ -75,19 +75,19 @@ public class Cambio implements Serializable {
         this.to = to;
     }
 
-    public BigDecimal getConversionFactor() {
+    public Double getConversionFactor() {
         return conversionFactor;
     }
 
-    public void setConversionFactor(BigDecimal conversionFactor) {
+    public void setConversionFactor(Double conversionFactor) {
         this.conversionFactor = conversionFactor;
     }
 
-    public BigDecimal getConvertedValue() {
+    public Double getConvertedValue() {
         return convertedValue;
     }
 
-    public void setConvertedValue(BigDecimal convertedValue) {
+    public void setConvertedValue(Double convertedValue) {
         this.convertedValue = convertedValue;
     }
